Stop using deprecated admission/v1beta1 api
