+++ admission/v1/errors.go
@@ -14,17 +14,17 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1beta1
+package v1
 
 import (
 	"net/http"
 
-	"k8s.io/api/admission/v1beta1"
+	admission "k8s.io/api/admission/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func StatusUninitialized() *v1beta1.AdmissionResponse {
-	status := &v1beta1.AdmissionResponse{}
+func StatusUninitialized() *admission.AdmissionResponse {
+	status := &admission.AdmissionResponse{}
 	status.Allowed = false
 	status.Result = &metav1.Status{
 		Status: metav1.StatusFailure, Code: http.StatusInternalServerError, Reason: metav1.StatusReasonInternalError,
@@ -33,8 +33,8 @@ func StatusUninitialized() *v1beta1.AdmissionResponse {
 	return status
 }
 
-func StatusInternalServerError(err error) *v1beta1.AdmissionResponse {
-	status := &v1beta1.AdmissionResponse{}
+func StatusInternalServerError(err error) *admission.AdmissionResponse {
+	status := &admission.AdmissionResponse{}
 	status.Allowed = false
 	status.Result = &metav1.Status{
 		Status: metav1.StatusFailure, Code: http.StatusInternalServerError, Reason: metav1.StatusReasonInternalError,
@@ -43,8 +43,8 @@ func StatusInternalServerError(err error) *v1beta1.AdmissionResponse {
 	return status
 }
 
-func StatusBadRequest(err error) *v1beta1.AdmissionResponse {
-	status := &v1beta1.AdmissionResponse{}
+func StatusBadRequest(err error) *admission.AdmissionResponse {
+	status := &admission.AdmissionResponse{}
 	status.Allowed = false
 	status.Result = &metav1.Status{
 		Status: metav1.StatusFailure, Code: http.StatusBadRequest, Reason: metav1.StatusReasonBadRequest,
@@ -53,8 +53,8 @@ func StatusBadRequest(err error) *v1beta1.AdmissionResponse {
 	return status
 }
 
-func StatusForbidden(err error) *v1beta1.AdmissionResponse {
-	status := &v1beta1.AdmissionResponse{}
+func StatusForbidden(err error) *admission.AdmissionResponse {
+	status := &admission.AdmissionResponse{}
 	status.Allowed = false
 	status.Result = &metav1.Status{
 		Status: metav1.StatusFailure, Code: http.StatusForbidden, Reason: metav1.StatusReasonForbidden,
