+++ registry/admissionreview/v1beta1/admission_review.go
@@ -19,7 +19,7 @@ package v1beta1
 import (
 	"context"
 
-	admission "k8s.io/api/admission/v1beta1"
+	admission "k8s.io/api/admission/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
